ethers constants bug fix

+++ contracts/contracts/Transactions.sol
@@ -5,9 +5,9 @@ pragma solidity ^0.8.9;
 // import "hardhat/console.sol";
 
 contract Transaction {
-    event Transfer(address sender, address receiver, uint8 amount, string message, string keyword,uint256 timestamp);
+    event Transfer(address sender, address receiver, uint amount, string message, string keyword,uint256 timestamp);
 
-   function send(address payable receiver, uint8 amount,string memory message, string memory keyword)public{
+   function send(address payable receiver, uint amount,string memory message, string memory keyword)public{
     emit Transfer(msg.sender,receiver, amount,message,keyword,block.timestamp);
    }
 }
